pred_pipeline/requirements.txt
astroid==2.15.4
black==23.3.0
boto3==1.26.90
botocore==1.29.129
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dill==0.3.6
exceptiongroup==1.1.1
fonttools==4.39.3
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
isort==5.12.0
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
matplotlib==3.7.1
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pathlib==1.0.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
pluggy==1.0.0
pylint==2.17.4
pyparsing==3.0.9
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.30.0
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.11.8
typing==3.7.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wrapt==1.15.0
zipp==3.15.0
